Add toggle to suppress KeyOrder element in maps
Introduce UclExportKeyOrder variable that allows to
disable adding the KeyOrder element to maps.

This allows to avoid errors when encoding/json.Unmarshal
is used to convert the map[string]interface{} output to
struct and the struct contains a map[string]string element.
Otherwise the following error is returned:

  json: cannot unmarshal array into Go value of type string

The KeyOrder element is exported by default to preserve
behaviour.